refactoring: request email flow
