+++ app/controllers/concerns/oauth_signed.rb
@@ -2,6 +2,16 @@ module OauthSigned
    extend ActiveSupport::Concern
 
    included do
+     def authorization_after_request_email
+       request.env['omniauth.auth'] = OmniAuth::AuthHash.new(
+         uid: params[:uid],
+         provider: params[:provider],
+         info: { email: params[:email] },
+         with_request_email: true
+       )
+
+       sign_in_flow
+     end
 
      private
