+++ config/routes.rb
@@ -8,7 +8,7 @@ Rails.application.routes.draw do
     get :sign_in, to: 'devise/sessions#new', as: :new_user_session
     post :sign_in, to: 'devise/sessions#create', as: :user_session
     delete :sign_out, to: 'devise/sessions#destroy', as: :destroy_user_session
-    post :continue_authorization, to: 'users/sessions#authorization_after_request_email'
+    post :continue_authorization, to: 'omniauth_callbacks#authorization_after_request_email'
   end
 
   concern :votable do |options|
